+++ Migration.lua
@@ -25,13 +25,13 @@ function Migration:Initialize()
     LedgerManager:RegisterOnUpdate(function(lag, uncommitted)
         if lag == 0 and uncommitted == 0 then
             if self.migrationOngoing then
-                LOG:Message("All migration entries were commited and executed. Congratulations!")
+                LOG:Message(CLM.L["All migration entries were commited and executed. Congratulations!"])
                 self.migrationOngoing = false
             end
             return
         end
         if self.migrationOngoing then
-            LOG:Message("Migration ongoing: %s(%s)", lag, uncommitted)
+            LOG:Message(CLM.L["Migration ongoing: %s(%s)"], lag, uncommitted)
         end
     end)
 end
@@ -44,11 +44,11 @@ local timestampCounter = {}
 function Migration:Migrate()
     if not ACL:CheckLevel(CONSTANTS.ACL.LEVEL.GUILD_MASTER) then return end
     if LedgerManager:Length() > 0 then
-        LOG:Message("Unable to execute migration. Entries already exist.")
+        LOG:Message(CLM.L["Unable to execute migration. Entries already exist."])
         return
     end
     Comms:Disable()
-    LOG:Message("Executing Addon Migration with comms disabled.")
+    LOG:Message(CLM.L["Executing Addon Migration with comms disabled."])
     -- self.timestamp = UTILS.GetCutoffTimestamp()
     self.playerCache = {}
     for i=1,GetNumGuildMembers() do
@@ -60,9 +60,9 @@ function Migration:Migrate()
     self:MigrateMonolithDKP()
     self:MigrateEssentialDKP()
     self:MigrateCommunityDKP()
-    LOG:Message("Migration complete. %s to apply and sync with others or go to %s to discard.",
+    LOG:Message(CLM.L["Migration complete. %s to apply and sync with others or go to %s to discard."],
         ColorCodeText("/reload", "00cc00"),
-        ColorCodeText("Minimap Icon -> Configuration -> Wipe events", "6699ff"))
+        ColorCodeText(CLM.L["Minimap Icon -> Configuration -> Wipe events"], "6699ff"))
 end
 
 function Migration:GetOldTimestampUnique()
@@ -88,7 +88,7 @@ end
 local function NewRoster(name)
     local timestamp = Migration:GetOldTimestampUnique()
     local roster = LEDGER_ROSTER.Create:new(timestamp, name, CONSTANTS.POINT_TYPE.DKP)
-    LOG:Message("New roster: [%s]", name)
+    LOG:Message(CLM.L["New roster: [%s]"], name)
     roster:setTime(timestamp)
     LedgerManager:Submit(roster)
     return name, timestamp
@@ -122,7 +122,7 @@ local function UpdatePoints(uid, targets, value)
 
     local t = entry:targets()
     if not t or (#t == 0) then
-        LOG:Message("UpdatePoints(): Empty targets list")
+        LOG:Message(CLM.L["UpdatePoints(): Empty targets list"])
         return
     end
 
@@ -149,7 +149,7 @@ local function ValidateAddon(addonName)
 	end
 
 	if not finished then
-		CLM:Error(addonName .. " has not finished loading!")
+		CLM:Error("%s has not finished loading!", addonName)
 		return false -- Should not happen
 	end
     return true
@@ -157,12 +157,12 @@ end
 
 function Migration:_MigrateMonolithEssential(addonName)
 	if not ValidateAddon(addonName) then
-        LOG:Message("Skipping " .. addonName)
+        LOG:Message(CLM.L["Skipping %s"], addonName)
         return
     end
     if not MonDKP_DKPTable then return end
     if not MonDKP_Loot then return end
-    LOG:Message("Migrating " .. addonName)
+    LOG:Message(CLM.L["Migrating %s"], addonName)
     -- Detect oldest timestamp
     local oldestTimestamp = GetServerTime()
     for _,entry in ipairs(MonDKP_Loot) do
@@ -175,7 +175,7 @@ function Migration:_MigrateMonolithEssential(addonName)
     end
     self.timestamp = oldestTimestamp - 86400 -- Since we increment it later on then we want to have a buffer
     -- Import profiles
-    LOG:Message("Importing %s entries from DKPTable", #MonDKP_DKPTable)
+    LOG:Message(CLM.L["Importing %s entries from DKPTable"], #MonDKP_DKPTable)
     self.playerDKP = {}
     self.playerList = {}
     local rosterCreated = false
@@ -201,11 +201,11 @@ function Migration:_MigrateMonolithEssential(addonName)
         end
     end
     if #self.playerList == 0 then
-        LOG:Error("Migration failure: Unable to create profiles")
+        LOG:Error(CLM.L["Migration failure: Unable to create profiles"])
         return
     end
     -- Add profiles to roster
-    LOG:Message("Adding %s profiles to %s", #self.playerList, rosterName)
+    LOG:Message(CLM.L["Adding %s profiles to %s"], #self.playerList, rosterName)
     AddProfilesToRoster(rosterUid, self.playerList)
     -- Import Loot History
     for _,entry in ipairs(MonDKP_Loot) do
@@ -224,7 +224,7 @@ function Migration:_MigrateMonolithEssential(addonName)
     for name,dkp in pairs(self.playerDKP) do
         UpdatePoints(rosterUid, GetPlayerGuid(self, name), dkp)
     end
-    LOG:Message("Import complete")
+    LOG:Message(CLM.L["Import complete"])
 end
 
 local function CommDKP_GetRealmName()
@@ -233,14 +233,14 @@ end
 
 function Migration:_MigrateCommunity()
     if not ValidateAddon("CommunityDKP") then
-        LOG:Message("Skipping CommunityDKP")
+        LOG:Message(CLM.L["Skipping CommunityDKP"])
         return
     end
     if not CommDKP_DB then return end
     if not CommDKP_DKPTable then return end
     if not CommDKP_Loot then return end
 
-    LOG:Message("Migrating CommunityDKP")
+    LOG:Message(CLM.L["Migrating %s"], "CommunityDKP")
 
     local realmFaction = CommDKP_GetRealmName()
     local guild = GetGuildInfo("player") or ""
@@ -263,7 +263,7 @@ function Migration:_MigrateCommunity()
         table.insert(teams, {id = id, name = teamInfo.name})
     end
     if #teams == 0 then
-        LOG:Error("Migration failure: Detected 0 teams")
+        LOG:Error(CLM.L["Migration failure: Detected 0 teams"])
         return
     end
     -- Detect oldest timestamp
@@ -286,7 +286,7 @@ function Migration:_MigrateCommunity()
     local playerProfiles = {}
     local teamProfiles = {}
     local teamRoster = {}
-    LOG:Message("Importing profiles from DKPTable")
+    LOG:Message(CLM.L["Importing profiles from DKPTable"])
     for _, team in ipairs(teams) do
         local rosterCreated = false
         local rosterName, rosterUid
@@ -309,22 +309,22 @@ function Migration:_MigrateCommunity()
                             NewProfile(GetPlayerGuid(self, name), name, class)
                             playerProfiles[name] = true
                             playerDKP[team.name][name] = dkp
-                            table.insert(teamProfiles[team.name], GetPlayerGuid(self, name))
                         end
+                        table.insert(teamProfiles[team.name], GetPlayerGuid(self, name))
                     end
                 end
             end
         end
     end
     if rawequal(next(playerProfiles), nil) then
-        LOG:Error("Migration failure: Unable to create profiles")
+        LOG:Error(CLM.L["Migration failure: Unable to create profiles"])
         return
     end
     -- Add profiles to rosters
     for _, team in ipairs(teams) do
         if DKPTable[team.id] then
             if #teamProfiles[team.name] > 0 then
-                LOG:Message("Adding %s profiles to %s", #teamProfiles[team.name], team.name)
+                LOG:Message(CLM.L["Adding %s profiles to %s"], #teamProfiles[team.name], team.name)
                 AddProfilesToRoster(teamRoster[team.name], teamProfiles[team.name])
             end
         end
@@ -346,7 +346,7 @@ function Migration:_MigrateCommunity()
                     end
                 end
             end
-            LOG:Message("Adding %s loot entries for team to %s", lootCount, team.name)
+            LOG:Message(CLM.L["Adding %s loot entries for team to %s"], lootCount, team.name)
         end
     end
     -- Set player DKP
@@ -356,9 +356,9 @@ function Migration:_MigrateCommunity()
             UpdatePoints(teamRoster[teamName], GetPlayerGuid(self, name), dkp)
             dkpSet = dkpSet + 1
         end
-        LOG:Message("Set DKP for %s players for team to %s", dkpSet, teamName)
+        LOG:Message(CLM.L["Set DKP for %s players for team to %s"], dkpSet, teamName)
     end
-    LOG:Message("Import complete")
+    LOG:Message(CLM.L["Import complete"])
 end
 
 function Migration:RegisterSlash()
@@ -366,7 +366,7 @@ function Migration:RegisterSlash()
         migrate = {
             type = "execute",
             name = "Migrate",
-            desc = "Execute migration from MonolithDKP, EssentialDKP or CommunityDKP",
+            desc = CLM.L["Execute migration from MonolithDKP, EssentialDKP or CommunityDKP"],
             handler = self,
             func = "Migrate",
         }