+++ Modules/AutoAwardManager/EncounterIDs.lua
@@ -1,129 +1,129 @@
 local _, CLM = ...
 local EncounterIDsMap = {
-    [1084] = "Onyxia",
-    [663] = "Lucifron",
-    [664] = "Magmadar",
-    [665] = "Gehennas",
-    [666] = "Garr",
-    [668] = "Baron Geddon",
-    [667] = "Shazzrah",
-    [669] = "Sulfuron Harbinger",
-    [670] = "Golemagg the Incinerator",
-    [671] = "Majordomo Executus",
-    [672] = "Ragnaros",
-    [610] = "Razorgore the Untamed",
-    [611] = "Vaelastrasz the Corrupt",
-    [612] = "Broodlord Lashlayer",
-    [613] = "Firemaw",
-    [614] = "Ebonroc",
-    [615] = "Flamegor",
-    [616] = "Chromaggus",
-    [617] = "Nefarian",
-    [709] = "The Prophet Skeram",
-    [710] = "Silithid Royalty",
-    [711] = "Battleguard Sartura",
-    [712] = "Fankriss the Unyielding",
-    [713] = "Viscidus",
-    [714] = "Princess Huhuran",
-    [715] = "Twin Emperors",
-    [716] = "Ouro",
-    [717] = "C'Thun",
-    [1107] = "Anub'Rekhan",
-    [1110] = "Grand Widow Faerlina",
-    [1116] = "Maexxna",
-    [1118] = "Patchwerk",
-    [1111] = "Grobbulus",
-    [1108] = "Gluth",
-    [1120] = "Thaddius",
-    [1117] = "Noth the Plaguebringer",
-    [1112] = "Heigan the Unclean",
-    [1115] = "Loatheb",
-    [1113] = "Instructor Razuvious",
-    [1109] = "Gothik the Harvester",
-    [1121] = "The Four Horsemen",
-    [1119] = "Sapphiron",
-    [1114] = "Kel'Thuzad",
-    [789] = "High Priest Thekal",
-    [784] = "High Priest Venoxis",
-    [791] = "High Priestess Arlokk",
-    [785] = "High Priestess Jeklik",
-    [786] = "High Priestess Mar'li",
-    [787] = "Bloodlord Mandokir",
-    [792] = "Jin'do the Hexxer",
-    [793] = "Hakkar",
-    [788] = "Edge of Madness",
-    [790] = "Gahz'ranka",
-    [718] = "Kurinnaxx",
-    [719] = "General Rajaxx",
-    [720] = "Moam",
-    [721] = "Buru the Gorger",
-    [722] = "Ayamiss the Hunter",
-    [723] = "Ossirian the Unscarred",
-    [652] = "Attumen the Huntsman", -- "Attumen the Huntsman",
-    [653] = "Moroes", -- "Moroes",
-    [654] = "Maiden of Virtue", -- "Maiden of Virtue",
-    [655] = "Opera Hall", -- "Opera Hall",
-    [656] = "The Curator", -- "The Curator",
-    [657] = "Terestian Illhoof", -- "Terestian Illhoof",
-    [658] = "Shade of Aran", -- "Shade of Aran",
-    [659] = "Netherspite", -- "Netherspite",
-    [660] = "Chess Event", -- "Chess Event",
-    [661] = "Prince Malchezaar", -- "Prince Malchezaar",
-    [662] = "Nightbane", -- "Nightbane"
-    [649] = "High King Maulgar", -- "High King Maulgar",
-    [650] = "Gruul the Dragonkiller", -- "Gruul the Dragonkiller"
-    [651] = "Magtheridon", -- "Magtheridon"
-    [623] = "Hydross the Unstable", -- "Hydross the Unstable",
-    [624] = "The Lurker Below", -- "The Lurker Below",
-    [625] = "Leotheras the Blind", -- "Leotheras the Blind",
-    [626] = "Fathom-Lord Karathress", -- "Fathom-Lord Karathress",
-    [627] = "Morogrim Tidewalker", -- "Morogrim Tidewalker",
-    [628] = "Lady Vashj", -- "Lady Vashj"
-    [730] = "Al'ar", -- "Al'ar",
-    [731] = "Void Reaver", -- "Void Reaver",
-    [732] = "High Astromancer Solarian", -- "High Astromancer Solarian",
-    [733] = "Kael'thas Sunstrider", -- "Kael'thas Sunstrider"
-    [601] = "High Warlord Naj'entus", -- High Warlord Naj'entus,
-    [602] = "Supremus", -- Supremus
-    [603] = "Shade of Akama", -- Shade of Akama
-    [604] = "Teron Gorefiend", -- Teron Gorefiend,
-    [605] = "Gurtogg Bloodboil", -- Gurtogg Bloodboil
-    [606] = "Reliquary of Souls", -- Reliquary of Souls
-    [607] = "Mother Shahraz", -- Mother Shahraz
-    [608] = "The Illidari Council", -- The Illidari Council
-    [609] = "Illidan Stormrage", -- Illidan Stormrage
-    [618] = "Rage Winterchill", -- High Warlord Naj'entus,
-    [619] = "Anetheron", -- Supremus
-    [620] = "Kaz'rogal", -- Shade of Akama
-    [621] = "Azgalor", -- Teron Gorefiend,
-    [622] = "Archimonde", -- Gurtogg Bloodboil
-    [1189] = "Akil'zon", -- "Akil'zon"
-    [1190] = "Nalorakk", -- Nalorakk
-    [1191] = "Jan'alai", -- Jan'alai
-    [1192] = "Halazzi", -- Halazzi,
-    [1193] = "Hex Lord Malacrass", -- Hex Lord Malacrass
-    [1194] = "Daakara", -- Daakara
-    [724] = "Kalecgos", -- "Kalecgos",
-    [725] = "Brutallus", -- "Brutallus",
-    [726] = "Felmyst", -- "Felmyst",
-    [727] = "Eredar Twins", -- "Eredar Twins",
-    [728] = "M'uru", -- "M'uru",
-    [729] = "Kil'jaeden", -- "Kil'jaeden"
+    [1084] = CLM.L["Onyxia"],
+    [663] = CLM.L["Lucifron"],
+    [664] = CLM.L["Magmadar"],
+    [665] = CLM.L["Gehennas"],
+    [666] = CLM.L["Garr"],
+    [668] = CLM.L["Baron Geddon"],
+    [667] = CLM.L["Shazzrah"],
+    [669] = CLM.L["Sulfuron Harbinger"],
+    [670] = CLM.L["Golemagg the Incinerator"],
+    [671] = CLM.L["Majordomo Executus"],
+    [672] = CLM.L["Ragnaros"],
+    [610] = CLM.L["Razorgore the Untamed"],
+    [611] = CLM.L["Vaelastrasz the Corrupt"],
+    [612] = CLM.L["Broodlord Lashlayer"],
+    [613] = CLM.L["Firemaw"],
+    [614] = CLM.L["Ebonroc"],
+    [615] = CLM.L["Flamegor"],
+    [616] = CLM.L["Chromaggus"],
+    [617] = CLM.L["Nefarian"],
+    [709] = CLM.L["The Prophet Skeram"],
+    [710] = CLM.L["Silithid Royalty"],
+    [711] = CLM.L["Battleguard Sartura"],
+    [712] = CLM.L["Fankriss the Unyielding"],
+    [713] = CLM.L["Viscidus"],
+    [714] = CLM.L["Princess Huhuran"],
+    [715] = CLM.L["Twin Emperors"],
+    [716] = CLM.L["Ouro"],
+    [717] = CLM.L["C'Thun"],
+    [1107] = CLM.L["Anub'Rekhan"],
+    [1110] = CLM.L["Grand Widow Faerlina"],
+    [1116] = CLM.L["Maexxna"],
+    [1118] = CLM.L["Patchwerk"],
+    [1111] = CLM.L["Grobbulus"],
+    [1108] = CLM.L["Gluth"],
+    [1120] = CLM.L["Thaddius"],
+    [1117] = CLM.L["Noth the Plaguebringer"],
+    [1112] = CLM.L["Heigan the Unclean"],
+    [1115] = CLM.L["Loatheb"],
+    [1113] = CLM.L["Instructor Razuvious"],
+    [1109] = CLM.L["Gothik the Harvester"],
+    [1121] = CLM.L["The Four Horsemen"],
+    [1119] = CLM.L["Sapphiron"],
+    [1114] = CLM.L["Kel'Thuzad"],
+    [789] = CLM.L["High Priest Thekal"],
+    [784] = CLM.L["High Priest Venoxis"],
+    [791] = CLM.L["High Priestess Arlokk"],
+    [785] = CLM.L["High Priestess Jeklik"],
+    [786] = CLM.L["High Priestess Mar'li"],
+    [787] = CLM.L["Bloodlord Mandokir"],
+    [792] = CLM.L["Jin'do the Hexxer"],
+    [793] = CLM.L["Hakkar"],
+    [788] = CLM.L["Edge of Madness"],
+    [790] = CLM.L["Gahz'ranka"],
+    [718] = CLM.L["Kurinnaxx"],
+    [719] = CLM.L["General Rajaxx"],
+    [720] = CLM.L["Moam"],
+    [721] = CLM.L["Buru the Gorger"],
+    [722] = CLM.L["Ayamiss the Hunter"],
+    [723] = CLM.L["Ossirian the Unscarred"],
+    [652] = CLM.L["Attumen the Huntsman"],
+    [653] = CLM.L["Moroes"],
+    [654] = CLM.L["Maiden of Virtue"],
+    [655] = CLM.L["Opera Hall"],
+    [656] = CLM.L["The Curator"],
+    [657] = CLM.L["Terestian Illhoof"],
+    [658] = CLM.L["Shade of Aran"],
+    [659] = CLM.L["Netherspite"],
+    [660] = CLM.L["Chess Event"],
+    [661] = CLM.L["Prince Malchezaar"],
+    [662] = CLM.L["Nightbane"],
+    [649] = CLM.L["High King Maulgar"],
+    [650] = CLM.L["Gruul the Dragonkiller"],
+    [651] = CLM.L["Magtheridon"],
+    [623] = CLM.L["Hydross the Unstable"],
+    [624] = CLM.L["The Lurker Below"],
+    [625] = CLM.L["Leotheras the Blind"],
+    [626] = CLM.L["Fathom-Lord Karathress"],
+    [627] = CLM.L["Morogrim Tidewalker"],
+    [628] = CLM.L["Lady Vashj"],
+    [730] = CLM.L["Al'ar"],
+    [731] = CLM.L["Void Reaver"],
+    [732] = CLM.L["High Astromancer Solarian"],
+    [733] = CLM.L["Kael'thas Sunstrider"],
+    [601] = CLM.L["High Warlord Naj'entus"],
+    [602] = CLM.L["Supremus"],
+    [603] = CLM.L["Shade of Akama"],
+    [604] = CLM.L["Teron Gorefiend"],
+    [605] = CLM.L["Gurtogg Bloodboil"],
+    [606] = CLM.L["Reliquary of Souls"],
+    [607] = CLM.L["Mother Shahraz"],
+    [608] = CLM.L["The Illidari Council"],
+    [609] = CLM.L["Illidan Stormrage"],
+    [618] = CLM.L["Rage Winterchill"],
+    [619] = CLM.L["Anetheron"],
+    [620] = CLM.L["Kaz'rogal"],
+    [621] = CLM.L["Azgalor"],
+    [622] = CLM.L["Archimonde"],
+    [1189] = CLM.L["Akil'zon"],
+    [1190] = CLM.L["Nalorakk"],
+    [1191] = CLM.L["Jan'alai"],
+    [1192] = CLM.L["Halazzi"],
+    [1193] = CLM.L["Hex Lord Malacrass"],
+    [1194] = CLM.L["Daakara"],
+    [724] = CLM.L["Kalecgos"],
+    [725] = CLM.L["Brutallus"],
+    [726] = CLM.L["Felmyst"],
+    [727] = CLM.L["Eredar Twins"],
+    [728] = CLM.L["M'uru"],
+    [729] = CLM.L["Kil'jaeden"],
 }
 CLM.EncounterIDsMap = EncounterIDsMap
 
 CLM.EncounterIDs = {}
 CLM.EncounterIDs.Classic = {
-    -- Onyxia
+
     {
-        name = "Onyxia's Lair",
+        name = CLM.L["Onyxia's Lair"],
         data = {
 	        { id = 1084, order =  1, name = EncounterIDsMap[1084] },
         },
     },
-    -- MC
+
     {
-        name = "Molten Core",
+        name = CLM.L["Molten Core"],
         data = {
             { id = 663, name = EncounterIDsMap[663] },
             { id = 664, name = EncounterIDsMap[664] },
@@ -137,9 +137,9 @@ CLM.EncounterIDs.Classic = {
             { id = 672, name = EncounterIDsMap[672] },
         },
     },
-    -- BWL
+
     {
-        name = "Blackwing Lair",
+        name = CLM.L["Blackwing Lair"],
         data = {
             { id = 610, name = EncounterIDsMap[610] },
             { id = 611, name = EncounterIDsMap[611] },
@@ -151,9 +151,9 @@ CLM.EncounterIDs.Classic = {
             { id = 617, name = EncounterIDsMap[617] },
         },
     },
-    -- AQ40
+
     {
-        name = "Temple of Ahn'Qiraj",
+        name = CLM.L["Temple of Ahn'Qiraj"],
         data = {
             { id = 709, name = EncounterIDsMap[709] },
             { id = 710, name = EncounterIDsMap[710] },
@@ -166,9 +166,9 @@ CLM.EncounterIDs.Classic = {
             { id = 717, name = EncounterIDsMap[717] },
         },
     },
-    -- Naxx
+
     {
-        name = "Naxxramas",
+        name = CLM.L["Naxxramas"],
         data = {
             { id = 1107, name = EncounterIDsMap[1107] },
             { id = 1110, name = EncounterIDsMap[1110] },
@@ -187,9 +187,9 @@ CLM.EncounterIDs.Classic = {
             { id = 1114, name = EncounterIDsMap[1114] },
         },
     },
-    -- ZG
+
     {
-        name = "Zul'Gurub",
+        name = CLM.L["Zul'Gurub"],
         data = {
             { id = 789, name = EncounterIDsMap[789] },
             { id = 784, name = EncounterIDsMap[784] },
@@ -203,9 +203,9 @@ CLM.EncounterIDs.Classic = {
             { id = 790, name = EncounterIDsMap[790] },
         },
     },
-    -- AQ20
+
     {
-        name = "Ruins of Ahn'Qiraj",
+        name = CLM.L["Ruins of Ahn'Qiraj"],
         data = {
             { id = 718, name = EncounterIDsMap[718] },
             { id = 719, name = EncounterIDsMap[719] },
@@ -218,108 +218,108 @@ CLM.EncounterIDs.Classic = {
 }
 
 CLM.EncounterIDs.TBC = {
-    -- Karazhan
+
     {
-        name = "Karazhan",
+        name = CLM.L["Karazhan"],
         data = {
-            { id = 652, name = EncounterIDsMap[652] }, -- "Attumen the Huntsman",
-            { id = 653, name = EncounterIDsMap[653] }, -- "Moroes",
-            { id = 654, name = EncounterIDsMap[654] }, -- "Maiden of Virtue",
-            { id = 655, name = EncounterIDsMap[655] }, -- "Opera Hall",
-            { id = 656, name = EncounterIDsMap[656] }, -- "The Curator",
-            { id = 657, name = EncounterIDsMap[657] }, -- "Terestian Illhoof",
-            { id = 658, name = EncounterIDsMap[658] }, -- "Shade of Aran",
-            { id = 659, name = EncounterIDsMap[659] }, -- "Netherspite",
-            { id = 660, name = EncounterIDsMap[660] }, -- "Chess Event",
-            { id = 661, name = EncounterIDsMap[661] }, -- "Prince Malchezaar",
-            { id = 662, name = EncounterIDsMap[662] }, -- "Nightbane"
+            { id = 652, name = EncounterIDsMap[652] },
+            { id = 653, name = EncounterIDsMap[653] },
+            { id = 654, name = EncounterIDsMap[654] },
+            { id = 655, name = EncounterIDsMap[655] },
+            { id = 656, name = EncounterIDsMap[656] },
+            { id = 657, name = EncounterIDsMap[657] },
+            { id = 658, name = EncounterIDsMap[658] },
+            { id = 659, name = EncounterIDsMap[659] },
+            { id = 660, name = EncounterIDsMap[660] },
+            { id = 661, name = EncounterIDsMap[661] },
+            { id = 662, name = EncounterIDsMap[662] },
         }
     },
-    -- Gruul's Lair
+
     {
-        name = "Gruul's Lair",
+        name = CLM.L["Gruul's Lair"],
         data = {
-            { id = 649, name = EncounterIDsMap[649] }, -- "High King Maulgar",
-            { id = 650, name = EncounterIDsMap[650] }, -- "Gruul the Dragonkiller"
+            { id = 649, name = EncounterIDsMap[649] },
+            { id = 650, name = EncounterIDsMap[650] },
         }
     },
-    --  Magtheridon
+
     {
-        name = "Magtheridon",
+        name = CLM.L["Magtheridon"],
         data = {
-            { id = 651, name = EncounterIDsMap[651] }, -- "Magtheridon"
+            { id = 651, name = EncounterIDsMap[651] },
         }
     },
-    -- SSC
+
     {
-        name = "Serpentshrine Cavern",
+        name = CLM.L["Serpentshrine Cavern"],
         data = {
-            { id = 623, name = EncounterIDsMap[623] }, -- "Hydross the Unstable",
-            { id = 624, name = EncounterIDsMap[624] }, -- "The Lurker Below",
-            { id = 625, name = EncounterIDsMap[625] }, -- "Leotheras the Blind",
-            { id = 626, name = EncounterIDsMap[626] }, -- "Fathom-Lord Karathress",
-            { id = 627, name = EncounterIDsMap[627] }, -- "Morogrim Tidewalker",
-            { id = 628, name = EncounterIDsMap[628] }, -- "Lady Vashj"
+            { id = 623, name = EncounterIDsMap[623] },
+            { id = 624, name = EncounterIDsMap[624] },
+            { id = 625, name = EncounterIDsMap[625] },
+            { id = 626, name = EncounterIDsMap[626] },
+            { id = 627, name = EncounterIDsMap[627] },
+            { id = 628, name = EncounterIDsMap[628] },
         }
     },
-    -- TK
+
     {
-        name = "Tempest Keep",
+        name = CLM.L["Tempest Keep"],
         data = {
-            { id = 730, name = EncounterIDsMap[730] }, -- "Al'ar",
-            { id = 731, name = EncounterIDsMap[731] }, -- "Void Reaver",
-            { id = 732, name = EncounterIDsMap[732] }, -- "High Astromancer Solarian",
-            { id = 733, name = EncounterIDsMap[733] }, -- "Kael'thas Sunstrider"
+            { id = 730, name = EncounterIDsMap[730] },
+            { id = 731, name = EncounterIDsMap[731] },
+            { id = 732, name = EncounterIDsMap[732] },
+            { id = 733, name = EncounterIDsMap[733] },
         }
     },
-    -- Black Temple
+
     {
-        name = "Black Temple",
+        name = CLM.L["Black Temple"],
         data = {
-            { id = 601, name = EncounterIDsMap[601] }, -- High Warlord Naj'entus,
-            { id = 602, name = EncounterIDsMap[602] }, -- Supremus
-            { id = 603, name = EncounterIDsMap[603] }, -- Shade of Akama
-            { id = 604, name = EncounterIDsMap[604] }, -- Teron Gorefiend,
-            { id = 605, name = EncounterIDsMap[605] }, -- Gurtogg Bloodboil
-            { id = 606, name = EncounterIDsMap[606] }, -- Reliquary of Souls
-            { id = 607, name = EncounterIDsMap[607] }, -- Mother Shahraz
-            { id = 608, name = EncounterIDsMap[608] }, -- The Illidari Council
-            { id = 609, name = EncounterIDsMap[609] }, -- Illidan Stormrage
+            { id = 601, name = EncounterIDsMap[601] },
+            { id = 602, name = EncounterIDsMap[602] },
+            { id = 603, name = EncounterIDsMap[603] },
+            { id = 604, name = EncounterIDsMap[604] },
+            { id = 605, name = EncounterIDsMap[605] },
+            { id = 606, name = EncounterIDsMap[606] },
+            { id = 607, name = EncounterIDsMap[607] },
+            { id = 608, name = EncounterIDsMap[608] },
+            { id = 609, name = EncounterIDsMap[609] },
         }
     },
-    -- Hyjal Summit
+
     {
-        name = "Hyjal Summit",
+        name = CLM.L["Hyjal Summit"],
         data = {
-            { id = 618, name = EncounterIDsMap[618] }, -- High Warlord Naj'entus,
-            { id = 619, name = EncounterIDsMap[619] }, -- Supremus
-            { id = 620, name = EncounterIDsMap[620] }, -- Shade of Akama
-            { id = 621, name = EncounterIDsMap[621] }, -- Teron Gorefiend,
-            { id = 622, name = EncounterIDsMap[622] }, -- Gurtogg Bloodboil
+            { id = 618, name = EncounterIDsMap[618] },
+            { id = 619, name = EncounterIDsMap[619] },
+            { id = 620, name = EncounterIDsMap[620] },
+            { id = 621, name = EncounterIDsMap[621] },
+            { id = 622, name = EncounterIDsMap[622] },
         }
     },
-    -- Zul'Aman
+
     {
-        name = "Zul'Aman",
+        name = CLM.L["Zul'Aman"],
         data = {
-            { id = 1189, name = EncounterIDsMap[1189] }, -- "Akil'zon"
-            { id = 1190, name = EncounterIDsMap[1190] }, -- Nalorakk
-            { id = 1191, name = EncounterIDsMap[1191] }, -- Jan'alai
-            { id = 1192, name = EncounterIDsMap[1192] }, -- Halazzi,
-            { id = 1193, name = EncounterIDsMap[1193] }, -- Hex Lord Malacrass
-            { id = 1194, name = EncounterIDsMap[1194] }, -- Daakara
+            { id = 1189, name = EncounterIDsMap[1189] },
+            { id = 1190, name = EncounterIDsMap[1190] },
+            { id = 1191, name = EncounterIDsMap[1191] },
+            { id = 1192, name = EncounterIDsMap[1192] },
+            { id = 1193, name = EncounterIDsMap[1193] },
+            { id = 1194, name = EncounterIDsMap[1194] },
         }
     },
-    -- Sunwell
+
     {
-        name = "Sunwell Plateau",
+        name = CLM.L["Sunwell Plateau"],
         data = {
-            { id = 724, name = EncounterIDsMap[724] }, -- "Kalecgos",
-            { id = 725, name = EncounterIDsMap[725] }, -- "Brutallus",
-            { id = 726, name = EncounterIDsMap[726] }, -- "Felmyst",
-            { id = 727, name = EncounterIDsMap[727] }, -- "Eredar Twins",
-            { id = 728, name = EncounterIDsMap[728] }, -- "M'uru",
-            { id = 729, name = EncounterIDsMap[729] } -- "Kil'jaeden"
+            { id = 724, name = EncounterIDsMap[724] },
+            { id = 725, name = EncounterIDsMap[725] },
+            { id = 726, name = EncounterIDsMap[726] },
+            { id = 727, name = EncounterIDsMap[727] },
+            { id = 728, name = EncounterIDsMap[728] },
+            { id = 729, name = EncounterIDsMap[729] }
         }
     }
 }