+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Classic Loot Manager v1.1.3
+## Fixes
+### Migration
+CommunityDKP migration should now properly add players to multiple teams.
+### Loot award with multiple linked alts
+Points should now be properly distributed when awarding loot to alt, whose main has multiple alts.
+### Profile creation
+You can now create profiles from guild by name through `/clm addprofile NAME`.
 # Classic Loot Manager v1.1.2
 ## Fixes
 ### Migration